refactor: extract booking broadcast helper
